Got initial batch of problems set up.
Seems like most of these are super easy. Many could be one-liners or generalized. They're also very repetitive. No23 and No14 are essentially the same. Only14 is a modification of that, essentially -- or vice versa. I may have to supplement these with others on a different site they recommended. Regardless, I plan on TDDing every one of them, even if they are stupendously easy.